requirements.txt
# Основные
fastapi==0.104.1
uvicorn[standard]==0.24.0

# База данных
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Pydantic v2
pydantic==2.5.0
pydantic-settings==2.1.0  # Добавляем!

# Аутентификация
python-jose[cryptography]==3.3.0
bcrypt==4.0.1

# Конфигурация
python-dotenv==1.0.0

# Утилиты
httpx==0.25.1

# Веб-интерфейс
jinja2==3.1.2
python-multipart==0.0.6
requests==2.31.0

# Для email валидации
email-validator==2.0.0

passlib
